Moved Coin Reduction to Internal and execute it with lock
